requirements.txt
flask
requests
yt-dlp
mutagen
python-dotenv
gunicorn
